More work to AlarmGenerationForm.
Splitted the AlarmXmlGenerator same as for IOXmlGenerator.
Changed how columns are handled in the GridHandler so now you have to add columns manually so it can also add comboBox where you prefer while maintaining string as a type in the binding.
Made drag only for text boxes.
ProjectSave and Setting are working fine for Alarm now.

TO-DO - Configuration and preview