Improved ProgramSettings. Added AutoSaveHandler.
Added an auto save handler created in the main form and passed to other forms.

Want to move all the settings inside a single class and pass them as parameters. Also need to improve the comparison and creation of snapshot to check for sub objects and count for them (This way just adding a new class to the settings will be counted to be saved if changed!)