More implementation to SimaticML.
Added more implementation to simatic ml. Changed how ID and UId are updated so there SHOULD be less confusion and spaghetti (Master class for locals that updates all his childrens).

Added Call, CallInfo and CallParameter so now they are updated when loading an xml file.

Change the SimaticDataType from enum to a class with static elements in it. Easier to handle and mantain.

Also refractored XMLClasses in their folder and updated to support functions descripted above.