Improved Suggestions.
A lot of improvements to the Suggestion Column (Now is very usable even without using the mouse!).
Added suggestions import from .xml file to be used in the IO Generation grid view.

Removed the IOData DbName column, added a MerkerAddress column, added new values for variable table start address in the config (Differentiate between input and output) and changed the logic on the xml is generated.

Fixed a bug where cancelling a file dialog would still try to save in an empty path causing an exception.