Refractoring in GridHandler. Improve Copy&Paste.
Removed the gridViews from the designers and now create it directly inside the GridHandler and adding it manually.

Done some changes around key combination and Copy&Paste as excel to avoid strange behaviours and to account invisible columns.

Also remove .Refresh() from the GridHandler.ChangeCell function to improve performance (Since is not needed there!). That method needs to be improved because while copying around 300 - 400 lines it takes good 4-5 seconds to complete everything. This should be done changing directly the DataSource instead of messing with cells.