Cleanup & Refractoring. Started work Localization.
Also done some improvements to the SuggestionColumn to avoid having to use the mouse while selected items inside the DropDown (Less stress to hands is always better).

Also done some changes to Localization and started adding it to the various part of the software. Need to write a little application to verify localization files to check for missing lines.

Changed the AutoSave time from an enum to an integer (Why i made an enum? Don't remember, don't care).