Big refactoring.
Added two more columns for AlarmGeneration to allow to define for each alarm what type of coil to use.

Removed the second parameter for GridHandler, only IGridData is kept, to have easier to handle code and less OOP mess. Also decoupled the preview system from the IGridData themselves to be, hopefully, easier to mantain and to access.

Refactored the GenerationPlaceholderHandler, to be hopefully easier to handle (Got it? Is a handler ehehehe)

Improved also XmlGeneration with many bug fixes.