Revert Localization to .resx files.
Added dirty state to GridScriptContainer and ScriptInfo (and included that state in the dirty chain of modules).

Save the program settings before asking for restarting when changing language.

Changed GridScript one line execution to be based on the currently selected row on the Grid instead of an internal index.

Added localization to some missing fields and renamed some wrong fileds inside Alarm configuration.