Major rework for GridScript. Added ObservableObject for ConfigForm.
Done a major rework to allow the GridScript form to be "universal". Before you would need to close and open a new one every time for different grids, now is always the same just with different binded GridHandlers. This way is MUCH more intuitive and easier to use.
While i was at it, i refactored the code to be a bit simpler and less repetitive.
Also made the save directly for the GridScript instread for the GridScriptContainer that i have removed (Unfortunately, saving break. Not worrying too much right now since is still just basically used by myself).

Also added a ObservableObject class that, used in the ConfigForm, allows me to change label text after i have already opened them.