Added more to SimaticML API. Implemented Custom Variable to alarm generation.
SImaticML implemented branches to SimaticPart to allow to add branch (Before there was only OR). Also added the MovePart, to allow to use the block MOVE.

TiaUtilities added a new custom variables to alarm generation to have a costant that is moved inside a defined variable for various uses. (Implemented all the fields, configuration and columns).